Move AdsAccountSelectControl to ConnectAds folder.
